+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * [#663](https://github.com/clojure-emacs/cider-nrepl/pull/663): Fix non-recognized `recur` inside `case` in the debugger.
 * [#664](https://github.com/clojure-emacs/cider-nrepl/pull/664): Fix continue-all in conditional break.
+* [#665](https://github.com/clojure-emacs/cider-nrepl/pull/665): Fix form location in map literals.
 
 ## 0.23.0 (2019-01-18)