+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New Features
 
+## 0.24.0 (2020-02-14)
+
 ### Bugs fixed
 
 * [#663](https://github.com/clojure-emacs/cider-nrepl/pull/663): Fix non-recognized `recur` inside `case` in the debugger.